+++ skyrim64_test/src/patches/rendering/codegen.cpp
@@ -89,10 +89,10 @@ void CreateXbyakCodeBlock()
 PatchCodeGen::PatchCodeGen(const PatchEntry *Patch, uintptr_t Memory, size_t MemorySize) : CodeGenerator(MemorySize, (void *)Memory)
 {
 #if TLS_DEBUG_ENABLE
-	const uint32_t structMemberOffset = Patch->Offset;
+	const uintptr_t structMemberOffset = Patch->Offset;
 #else
 	// offsetof(IMAGE_TLS_DIRECTORY->StartAddressOfRawData, BSGraphicsRendererGlobals) + offsetof(BSGraphicsRendererGlobals, var)
-	const uint32_t structMemberOffset = BSGRAPHICS_TLS_BASE_OFFSET + Patch->Offset;
+	const uintptr_t structMemberOffset = BSGRAPHICS_TLS_BASE_OFFSET + Patch->Offset;
 #endif
 
 	auto& s = GetFreeScratch(Patch->Register, Patch->Base, Patch->Index);	// Scratch register
